NextChain experimental changes -- these changes are compressed into a single commit because rebasing multiple commits is extremely painful.

+++ src/validation/parallel.h
@@ -164,7 +164,7 @@ private:
         uint256 hashPrevBlock;
         uint32_t nChainWork; // chain work for this block.
         uint32_t nMostWorkOurFork; // most work for the chain we are on.
-        uint32_t nSequenceId;
+        uint64_t nSequenceId;
         int64_t nStartTime;
         uint64_t nBlockSize;
         bool fQuit;
